requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.4
PyYAML==6.0.2
python-dotenv==1.0.1
aiohttp==3.10.5
websockets==12.0
numpy==2.0.1
optuna==3.6.1
